app/requirements.txt
packaging==24.1
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-generativeai==0.7.2
PyPDF2==3.0.1
youtube-transcript-api==0.6.2
yt-dlp==2024.7.9
faiss-cpu==1.8.0
numpy==1.26.4
python-dotenv==1.0.1
python-multipart==0.0.9